ALINE/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.1",
    "pyyaml>=6.0.0",
    "torch>=2.5.0",
    "fastapi>=0.121.2",
    "uvicorn>=0.38.0",
    "pydantic>=2.12.4",
    "httpx>=0.27.0",
    "scipy>=1.16.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "requests>=2.32.5",
]

[tool.hatch.build.targets.wheel]
packages = ["models", "service"]
